+++ src/components/Comments.jsx
@@ -1,0 +1,3 @@
+export default function Comments (){
+    return (<h1>Home</h1>)
+}
