+++ src/components/modals/Comments.jsx
@@ -1,0 +1,5 @@
+export default function Comments() {
+    return (
+        <><h1>Comments</h1></>
+    )
+}
